+++ derive-examples/src/main.rs
@@ -5,6 +5,8 @@ mod custom_copy;
 #[cfg(test)]
 mod gtor_skip;
 #[cfg(test)]
+mod phantom;
+#[cfg(test)]
 mod stor_skip;
 
 fn main() {}